Update patch set 2
Patch Set 2:

Luca, can you check this?  The only way I could get the plugin to not NPE is by adding dummy values for the other settings.

And even after, I still can't get it to display anything.  Not sure if it's broken or I'm not using it properly...

Patch-set: 2